requirements.txt
# Core ML libraries
transformers==4.55.0
accelerate>=0.30
tokenizers>=0.19
peft>=0.11.0
safetensors

# PyTorch and vision
torch>=2.0.0
torchvision
pillow

# Quantization
bitsandbytes

# Data processing
datasets

# Cloud storage (for GCP)
gcsfs
google-cloud-storage
google-auth

# Utilities
rich
qwen_vl_utils